Moved tests to correct file
